Make wording more generic
With the intention of addressing https://github.com/heathd/twentyfiveten/issues/3

To satisfy other use cases, the prompts about bold ideas can be incorporated
into the challenge text. If we make glue text more generic it makes it more
flexible without having to build in loads of complicated logic. I've alluded to
this idea by pre-filling the challenge text box for new ideas with the classic
25/10 invitation.